+++ examples/clickhouse_integrations/kafka_source/variables.tf
@@ -1,0 +1,4 @@
+variable "aiven_api_token" {
+  description = "Aiven console API token"
+  type        = string
+}