+++ examples/clickhouse_integrations/postgres_source/project.tf
@@ -1,0 +1,4 @@
+// Sample project name
+resource "aiven_project" "clickhouse_postgres_source" {
+  project = "clickhouse-postgres-source"
+}