+++ CHANGELOG.md
@@ -9,6 +9,7 @@ nav_order: 1
 ## [MAJOR.MINOR.PATCH] - YYYY-MM-DD
 - Add ClickHouse examples:
   - Standalone service
+  - Integration with Kafka source
 
 ## [3.9.0] - 2022-12-01